sgpa calculator_1/requirements.txt
streamlit
tabula-py
pandas
PyPDF2